FormHelper specs, context cleanup
I added the start of some spec tests for the FormHelper. Also went
through and rephrased the 'context' and 'it' expressions to read more
naturally, like:

  WebHelper#should_see
    passes when
      String
        exists
        is within the scope
      Regexp
        matches
        matches within the scope
    fails when
      String
        does not exist
        is not within the scope
      Regexp
        does not match
        matches but is not within the scope

Eliminated words like "should" from the spec descriptions, since I
believe tests should be expressed in terms of asserted truths that can
be falsified by a failure, and "should" is not a falsifiable assertion.