Renamed _contain to _equal, removed questionable methods
Due to potential confusion between 'contain' and 'include', I renamed
'dropdown_should_contain' to 'dropdown_should_equal', to make it clearer
that we're looking for an exact match on the dropdown contents (as
opposed to 'dropdown_should_include', which looks for a value anywhere
in the dropdown's options).

Fixed the 'field_should_be_empty' method to raise an exception on failure,
instead of just returning false.

Also removed 'field_should_contain_date', since it depends on a piece of
code that was not included in this library, and 'fill_in_tinymce' since
it's a bit too specialized for a general-purpose form helper.