First steps toward webrat support
While taking my first steps toward adding support for Webrat in addition
to Capybara, I found a number of helpers that are identical for both
backends. Rather than duplicate all this code in both modules, I decided
to move these into lib/kelp/helper.rb, which is then included by both
the Capybara and Webrat modules.

I also refactored the page_should_have and page_should_not_have methods
so that the common code is in helper.rb, and the bits that are different
between the two platforms are separated into their respective modules.

I'm still not entirely comfortable with using modules for all the
namespacing--classes would be neater, I think.

The Capybara spec tests are all passing at this point, so it's a good
time to check in what's working, messy though it may be.