Cucumber-test proof of concept
This commit adds an initial proof of concept for testing Kelp's own step
definitions using Cucumber, executed against the same Sinatra app that
was already used by the spec-test suite. It is closely modeled after
email-spec's self-tests.

The Sinatra app now resides in examples/sinatra_app, which contains its
own features/ directory; this allows Cucumber and RSpec to share it. The
main features/ directory is now a basic single-scenario test that
invokes cucumber on the features within the Sinatra app.

Of course, Cucumber is now a development dependency (along with its
various dependencies). Cucumber-Sinatra does not appear to be needed
right now, although that may change.