Fix dropdown_should_equal
Due to Rails monkeypatching, the behavior of "String".first is different
from regular Ruby's "String".first; in Rails, only the first character
of the string is returned. Took out the use of "first", and added an
exception when attempting to use dropdown_should_equal with a
multiple-selection dropdown (which actually does have an array in
field.value).