should_see / should_not_see check multiple strings
Modified the should_see and should_not_see functions to check all of the
provided strings, and collect a list of the ones that do not meet
expectations; they then raise an exception if one or more expectations
were not met (instead of failing at the first unexpected result). Added
spec and cucumber tests to cover this contingency, along with a custom
Kelp::Unexpected exception (probably the first of many).