Fix wonky quoting behavior in 'within' clause
In order to support the `selector_for` method provided by recent
versions of cucumber-rails, the 'within' clause was silently stripping
off surrounding double-quotes. This caused some strange behavior when a
quoted identifier was passed; although it seemed to work, Cucumber's
output was omitting the initial " and duplicating the trailing
character. Thus this:

    Then I should see "Foo" within "#bar"

was outputted like this:

    Then I should see "Foo" within #barr"

Modifying the WITHIN regex to swallow any leading and trailing
double-quotes, and making the inner expression match anything except
double-quotes, allows these to be treated correctly. This makes it
unnecessary for the `within_scope` method to silently strip the quotes.