Ugly overhaul of fill-in methods
In order to raise more descriptive, Kelp-specific exceptions, some major
uglification of the `select_or_fill` method was needed. Part of this
difficulty is due to the fact that Capybara raises `ElementNotFound`
both when a field does not exist at all, and when a dropdown field
exists but does not include the expected value. An understandable state
of affairs; but rather than having Kelp mimic this ambiguity, new
exceptions `FieldNotFound` and `OptionNotFound` were added.

I suspect even more exception types will be needed in the near future
(especially to disambiguate the many uses of Kelp::Unexpected). It may
also become necessary to split the form-field-filling from the
form-field-verification methods, since field.rb is getting pretty
longish.